--- a/lib/JSON2CSVAsyncParser.js
+++ b/lib/JSON2CSVAsyncParser.js
@@ -2,6 +2,7 @@
 
 const { Transform } = require('stream');
 const JSON2CSVTransform = require('./JSON2CSVTransform');
+const { fastJoin } = require('./utils');
 
 class JSON2CSVAsyncParser {
   constructor(opts, transformOpts) {
@@ -40,10 +41,10 @@ class JSON2CSVAsyncParser {
 
   promise() {
     return new Promise((resolve, reject) => {
-      let csv = '';
+      let csvBuffer = [];
       this.processor
-        .on('data', chunk => (csv += chunk.toString()))
-        .on('finish', () => resolve(csv))
+        .on('data', chunk => csvBuffer.push(chunk.toString()))
+        .on('finish', () => resolve(fastJoin(csvBuffer, '')))
         .on('error', err => reject(err));
     });
   }