--- a/lib/JSON2CSVBase.js
+++ b/lib/JSON2CSVBase.js
@@ -2,7 +2,7 @@
 
 const os = require('os');
 const lodashGet = require('lodash.get');
-const { setProp, flattenReducer } = require('./utils');
+const { getProp, setProp, fastJoin, flattenReducer } = require('./utils');
 
 class JSON2CSVBase {
   constructor(opts) {
@@ -29,7 +29,7 @@ class JSON2CSVBase {
       : '"';
     processedOpts.doubleQuote = typeof processedOpts.doubleQuote === 'string'
       ? processedOpts.doubleQuote
-      : Array(3).join(processedOpts.quote);
+      : processedOpts.quote + processedOpts.quote;
     processedOpts.header = processedOpts.header !== false;
     processedOpts.includeEmptyRows = processedOpts.includeEmptyRows || false;
     processedOpts.withBOM = processedOpts.withBOM || false;
@@ -49,7 +49,9 @@ class JSON2CSVBase {
       if (typeof fieldInfo === 'string') {
         return {
           label: fieldInfo,
-          value: row => lodashGet(row, fieldInfo, this.opts.defaultValue),
+          value: (fieldInfo.includes('.') || fieldInfo.includes('['))
+            ? row => lodashGet(row, fieldInfo, this.opts.defaultValue)
+            : row => getProp(row, fieldInfo, this.opts.defaultValue),
           stringify: true,
         };
       }
@@ -62,7 +64,9 @@ class JSON2CSVBase {
         if (typeof fieldInfo.value === 'string') {
           return {
             label: fieldInfo.label || fieldInfo.value,
-            value: row => lodashGet(row, fieldInfo.value, defaultValue),
+            value: (fieldInfo.value.includes('.') || fieldInfo.value.includes('['))
+              ? row => lodashGet(row, fieldInfo.value, defaultValue)
+              : row => getProp(row, fieldInfo.value, defaultValue),
             stringify: fieldInfo.stringify !== undefined ? fieldInfo.stringify : true,
           };
         }
@@ -93,9 +97,10 @@ class JSON2CSVBase {
    * @returns {String} titles as a string
    */
   getHeader() {
-    return this.opts.fields
-      .map(fieldInfo => this.processValue(fieldInfo.label, true))
-      .join(this.opts.delimiter);
+    return fastJoin(
+      this.opts.fields.map(fieldInfo => this.processValue(fieldInfo.label, true)),
+      this.opts.delimiter
+    );
   }
 
   memoizePreprocessRow() {
@@ -139,15 +144,20 @@ class JSON2CSVBase {
    * @returns {String} CSV string (row)
    */
   processRow(row) {
-    if (!row
-        || (Object.getOwnPropertyNames(row).length === 0
-          && !this.opts.includeEmptyRows)) {
+    if (!row) {
       return undefined;
     }
 
-    return this.opts.fields
-      .map(fieldInfo => this.processCell(row, fieldInfo))
-      .join(this.opts.delimiter);
+    const processedRow = this.opts.fields.map(fieldInfo => this.processCell(row, fieldInfo));
+
+    if (!this.opts.includeEmptyRows && processedRow.every(field => field === undefined)) {
+      return undefined;
+    }
+
+    return fastJoin(
+      processedRow,
+      this.opts.delimiter
+    );
   }
 
   /**
@@ -173,61 +183,39 @@ class JSON2CSVBase {
       return undefined;
     }
 
-    const isValueString = typeof value === 'string';
-    if (isValueString) {
-      value = value
-        .replace(/\n/g, '\u2028')
-        .replace(/\r/g, '\u2029')
-        .replace(/\t/g, '\u21E5');
+    const valueType = typeof value;
+    if (valueType !== 'boolean' && valueType !== 'number' && valueType !== 'string') {
+      value = JSON.stringify(value);
+
+      if (value === undefined) {
+        return undefined;
+      }
+
+      if (value[0] === '"') {
+        value = value.replace(/^"(.+)"$/,'$1');
+      }
     }
 
-    //JSON.stringify('\\') results in a string with two backslash
-    //characters in it. I.e. '\\\\'.
-    let stringifiedValue = (stringify
-      ? JSON.stringify(value)
-      : value);
+    if (typeof value === 'string') {
+      if(value.includes(this.opts.quote)) {
+        value = value.replace(new RegExp(this.opts.quote, 'g'), this.opts.doubleQuote);
+      }
 
-    if (typeof value === 'object' && !/^".*"$/.test(stringifiedValue)) {
-      // Stringify object that are not stringified to a
-      // JSON string (like Date) to escape commas, quotes, etc.
-      stringifiedValue = JSON.stringify(stringifiedValue);
+      // This should probably be remove together with the whole strignify option
+      if (stringify) {
+        value = `${this.opts.quote}${value}${this.opts.quote}`;
+      } else {
+        value = value
+          .replace(new RegExp(`^${this.opts.doubleQuote}`), this.opts.quote)
+          .replace(new RegExp(`${this.opts.doubleQuote}$`), this.opts.quote);
+      }
+
+      if (this.opts.excelStrings) {
+        value = `"="${value}""`;
+      }
     }
 
-    if (stringifiedValue === undefined) {
-      return undefined;
-    }
-
-    if (isValueString) {
-      stringifiedValue = stringifiedValue
-        .replace(/\u2028/g, '\n')
-        .replace(/\u2029/g, '\r')
-        .replace(/\u21E5/g, '\t');
-    }
-
-    if (this.opts.quote === '"') {
-      // Replace automatically escaped single quotes by doubleQuotes
-      stringifiedValue = stringifiedValue
-        .replace(/\\"(?!$)/g, this.opts.doubleQuote);
-    } else {
-      // Unescape double quotes ('"')
-      // Replace single quote with double quote
-      // Replace wrapping quotes
-      stringifiedValue = stringifiedValue
-        .replace(/\\"(?!$)/g, '"')
-        .replace(new RegExp(this.opts.quote, 'g'), this.opts.doubleQuote)
-        .replace(/^"/, this.opts.quote)
-        .replace(/"$/, this.opts.quote);
-    }
-
-    // Remove double backslashes
-    stringifiedValue = stringifiedValue
-      .replace(/\\\\/g, '\\');
-
-    if (this.opts.excelStrings && typeof value === 'string') {
-      stringifiedValue = '"="' + stringifiedValue + '""';
-    }
-
-    return stringifiedValue;
+    return value;
   }
 
   /**