--- a/lib/JSON2CSVParser.js
+++ b/lib/JSON2CSVParser.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const JSON2CSVBase = require('./JSON2CSVBase');
-const { flattenReducer } = require('./utils');
+const { fastJoin, flattenReducer } = require('./utils');
 
 class JSON2CSVParser extends JSON2CSVBase {
   constructor(opts) {
@@ -73,10 +73,10 @@ class JSON2CSVParser extends JSON2CSVBase {
    * @returns {String} CSV string (body)
    */
   processData(data) {
-    return data
-      .map(row => this.processRow(row))
-      .filter(row => row) // Filter empty rows
-      .join(this.opts.eol);
+    return fastJoin(
+      data.map(row => this.processRow(row)).filter(row => row), // Filter empty rows
+      this.opts.eol
+    );
   }
 }