--- a/lib/utils.js
+++ b/lib/utils.js
@@ -1,5 +1,9 @@
 'use strict';
 
+function getProp(obj, path, defaultValue) {
+  return obj[path] === undefined ? defaultValue : obj[path];
+}
+
 function setProp(obj, path, value) {
   const pathArray = Array.isArray(path) ? path : path.split('.');
   const key = pathArray[0];
@@ -18,7 +22,25 @@ function flattenReducer(acc, arr) {
   }
 }
 
+function fastJoin(arr, separator) {
+  let isFirst = true;
+  return arr.reduce((acc, elem) => {
+    if (elem === null || elem === undefined) {
+      elem = '';
+    }
+
+    if (isFirst) {
+      isFirst = false;
+      return `${elem}`;
+    }
+
+    return `${acc}${separator}${elem}`;
+  }, '');
+}
+
 module.exports = {
+  getProp,
   setProp,
+  fastJoin,
   flattenReducer
 };