--- a/test/CLI.js
+++ b/test/CLI.js
@@ -560,8 +560,8 @@ module.exports = (testRunner, jsonFixtures, csvFixtures) => {
       const csv = stdout;
       t.equal(csv, [
       '"a string"',
-      '"with a \ndescription\\n and\na new line"',
-      '"with a \r\ndescription and\r\nanother new line"'
+      '"with a \u2028description\\n and\na new line"',
+      '"with a \u2029\u2028description and\r\nanother new line"'
     ].join('\r\n'));
       t.end();
     });